fix: unit tests

+++ utils/typeutils/resolver_test.go
@@ -130,7 +130,7 @@ func TestResolve(t *testing.T) {
 				dataType types.DataType
 				nullable bool
 			}{
-				"int_ptr":     {dataType: types.Int32, nullable: false},
+				"int_ptr":     {dataType: types.Int64, nullable: false},
 				"string_ptr":  {dataType: types.String, nullable: false},
 				"bool_ptr":    {dataType: types.Bool, nullable: false},
 				"float64_ptr": {dataType: types.Float64, nullable: false},
